Try to fix hardware loop
The loop_end pattern being a jump insn and having output operands
can't be reloaded correctly by the reload pass. A previous change was
introduced, inspired by arc's backend, but it was not correctly
implemented. When reload needs to reload the loop counter, it gives a
MEM rtx that is temporaly accepted by the pattern. This memory
location still needs to be stored in a register and wrote back when
finished: this is what this change does.

ref T7970

	gcc/
	* config/k1/k1.c (hwloop_fail): Detect case of loop_end that needs
	to be reloaded and emit corresponding mem write.
	(hwloop_pattern_reg): Match pattern with reload.
	* config/k1/k1.md (doloop_end): Add a scratch register used when
	reload needs to be handled.
	(loop_end): Convert to split. Split in case reload is needed.
	(loop_end_withreload): New.