Rework of label handling in PIC code
There was a mismatch between k1_legitimate_pic_operand_p,
k1_analyze_address and MD description for how LABELs are handled in
PIC code. From now on, a LABEL_REF is rejected as a PIC operand. It
should be handled early by the mov expander whose predicate accepts
LABEL_REF during PIC. It then emits an add_pcrel pattern that wraps
the LABEL_REF in an UNSPEC_PCREL. This UNSPEC_PCREL is then accepted
as an operand if the optimizer moves it around.
ref T9557

	gcc/
	* config/k1/k1.c (k1_analyze_address): Refactor the tests, fix the comments.
	(k1_legitimate_pic_operand_p): Reject LABEL_REF.
	(gen_add2_pcrel_insn): Helper for generating the correct pattern
	based on Pmode.
	(k1_expand_mov_constant): Emit add_pcrel pattern for LABEL_REF in PIC.
	(k1_output_addr_const_extra): Remove old comments, reindent using
	correct GNU style.
	* config/k1/k1.md (mov<mode>): Change predicate to accept
	LABEL_REF in PIC.
	(add_pcrel_<mode>): Refactor to make the PLUS and dep on pc explicit.
	* config/k1/predicates.md (k1_mov_operand): New. A general_operand
	that also accepts LABEL_REF during PIC (LEGITIMATE_PIC_OPERAND
	makes general_operand fail in this case).
	(symbolic_operand): New.