Make load packing a bit better and less constrained
Do not force load packing to be correctly aligned unless
-fstrict-align is used as in T9083

Better handling of pattern makes possible the matching of non-null
base offsets as seen in T10342.
adjust_address_nv won't emit a PLUS with a const0 which prenvented our
load_multiple patterns to match. Forcing the PLUS workaround this,
even if it is probably not the best solution.

	gcc/
	* config/k1/k1-protos.h (k1_expand_load_multiple): New.
	* config/k1/k1.c (k1_pack_load_store): Handle case where MEM has a
	REG instead of PLUS.
	(k1_expand_load_multiple): New, extracted from k1.md. Force a PLUS
	when adjust_address_nv would remove it because of the 0 constant.
	(k1_load_multiple_operation_p): Relax alignment constant.
	* config/k1/k1.md (load_multiple): Extract C and move to k1.c.
	(*lq_multiple_cached_m8): Removed, bug fixed.
	* testsuite/gcc.target/k1/lo_peephole.c: Add test for non 0 base
	offset.
	* testsuite/gcc.target/k1/lq_peephole.c: Likewise.

Differential Revision: D3046