Fix hardware loop handling
In case of failed hwloop, the mem store for the loop counter was not
inserted at the correct location, leading to it being placed after a
jump back. In some cases, the store ends up being the last insn of a
leaf BB, leading the insn to be bundled with whatever is following.

ref T12028

	gcc/
	* config/k1/k1.c (hwloop_fail): Emit the store before the loop_end
	insn.

Differential Revision: D3880