Fix incorrect handling of large argument in odd register
When moving to/from an incorrectly aligned register large values, gcc was trying
to split the operations. Spliting a memory access is not possible in the general
case and would need an extra register (too late for this). It would be also
better to not split the memory access and split its result instead.

This is happening before reload, so we can use a temp reg to do the access.

ref T15980

	gcc/
	* config/kvx/vector.md (mov<mode> ALL128): Use temp reg if needed.
	(mov<mode> ALL256): Likewise.
	(*mov<mode>_oddreg): Do not accept any memory.
	(*mov<mode>_misalign_reg): Likewise.
	* testsuite/gcc.target/kvx/t15980_largearg_oddreg_mem.c: New.
	* config/kvx/kvx.md (*mov_quad_oddreg): Do not accept any
	memory.memory

Differential Revision: D11609