Fix for small integer return values and enable Float16 mangling.
Summary:
When returning integers smaller that 8 bytes, the current KVX implementation
leaves the padding bits undefined. In most cases the padding is zero thanks to
the *W instructions that clear the 32 upper bits, but not in case a function
returns a specific element of a vector operated in that function.  This patch
ajusts GCC so that upper bits are cleared when returnning small integers.
Also, the missing C++ mangling string for the _Float16 type has been added.

  gcc/config/kvx/
  * kvx.c: add (kvx_promote_function_mode) to implement hook
  TARGET_PROMOTE_FUNCTION_MODE; in (kvx_function_value), apply the type
  promotion rules of (promote_function_mode); define
  TARGET_OMIT_STRUCT_RETURN_REG false; add (kvx_mangle_type) to implement hook
  TARGET_MANGLE_TYPE; for (kvx_mangle_type), apply the same mangling for 16-bit
  FP as in the aarch64.  * kvx.h: simplify implementation of (PROMOTE_MODE).
  gcc/testsuite/gcc.target/kvx/
  * lsu_3_sizes.c: removed as its scan-rtl-dump-times patterns are obsolete.

Test Plan: CI

Reviewers: jhascoet, O11 gcc!

Differential Revision: https://phab.kalray.eu/D15914