Enable assembly directive .nop for kvx
Summary:
Have to emit nop bundle and so:
     nop
     ;;

As ";;" is interpreted by start_line hook and replaced with pseudo op "be" to terminate the bundle
so we have to implement md_emit_single_noop hook.
It calls md_assemble with "nop" and with "be" to emit the bundle.
Without this support, gas complain about too many instruction in the bundle as md_assemble never
seen "be" pseudo op.

Test Plan: CI

Reviewers: O102 binutils!

Differential Revision: https://phab.kalray.eu/D16705