Enable symbols to be used as signed32 operands in 64-bit assembler.
Summary:
Symbols are assumed the pointer size by the assembler, so it does not allow
a symbol to be used in place of a 32-bit immediate in instruction operands
unless -m32 is passed. This patch introduces the '@signed32' pseudo-function
that enables to encode such symbolic operands in KVX 32-bit magic immediates.
This patch does not require update of information generated from MDS.

Test Plan: CI

Reviewers: jvillette, O102 binutils!

Differential Revision: https://phab.kalray.eu/D16762