rauc: Install RAUC to staging directory
Summary:
nvme-emulation needs to communicate with RAUC to implement fw-activate
command. The good way to do so is to use the D-Bus interface so we needs
to install rauc in the staging directory for other buildroot packages to
find the interface xml file.

Ref T18689

Test Plan: n/a

Maniphest Tasks: T18689

Differential Revision: https://phab.kalray.eu/D19799