libgcc: split all functions in their own sections
Summary:
In a size-restraint environment, including unused functions from a lib
is burdensome and leads to a lot of work to understand where the
superfluous code comes from, and to prune it afterwards.

This patch tries to alleviate this burden by using the dedicated flag
(already used on a lot of libs elsewhere in the toolchain) in libgcc.
The known issues that could arise are a significant increase in size of
the lib itself, longer build times, and potential debugging troubles.

Test Plan:
Running a .from thorough all of CI stages, plus local tests to estimate
the gain in various use cases.

Reviewers: O11 gcc!

Differential Revision: https://phab.kalray.eu/D22436