ltp-testsuite: remove unsupported tests in POST_CONFIG_HOOK
Summary:
Some tests are not supported by musl and
should be removed when using a musl toolchain.
Currently the removal of unsupported tests is done in
a POST_PATCH_HOOK. However since we use a local.mk,
patches are not applied and the patch hook is not executed.

This commit move the removal of unsupported tests in
a POST_CONFIG_HOOK.

Ref T19769

Test Plan: Build locally + CI

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T19769

Differential Revision: https://phab.kalray.eu/D22562