Improve support for 512-bit vectors and __builtin_convertvector().
Summary:
Ensure that the standard arithmetic SPNs on 512-bit vectors does not
falls back to operating the lanes in scalar mode. This is achieved by
ensuring the corresponding 256-bit vector SPNs exist. This does not work
for the vector conversions that change the bitdidth of lanes, so
specific patterns are added for use by __builtin_convertvector().
Also add patterns for the MADDWQ and MSBFWQ kv3-2 instructions.

ref T20514

  gcc/config/kvx/
  * builtin.md: move patterns "kvx_mul*", "kvx_madd*", "kvx_msbf*" from
  vector.md; improve "kvx_zx<widenx>_" to use the kv3-2 ZXL*, ZXM*; add
  pattern "kvx_fhighw" and use it in patterns "kvx_frecwp", "kvx_frsrwp".
  * iterators.md: add mode attribute QWIDE for the "trunc<wide><mode>2"
  and "extend<mode><wide>2" SPNs; add mode iterators VXHF and VXDF for
  refactoring add mode iterators V128J, V256G to support conversion SPNs;
  * vector.md: move patterns to builtin.md; add SPNs "float<mask><mode>2",
  "floatuns<mask><mode>2", "fix_trunc<mode><mask>2",
  "fixuns_trunc<mode><mask>2", "trunc<wide><mode>2", "extend<mode><wide>2";
  fix patterns using sxlbhq, sxmbhq, sxlhwp, sxmhwp to use alu_thin type.

Test Plan: CI

Reviewers: piannetta, O11 gcc!

Maniphest Tasks: T20514

Differential Revision: https://phab.kalray.eu/D26026