Fix wrong offset with two consecutive hwloops
Summary:
When there is two consecutive hardware loops, the address of the end of
the first loop ends up within the second hardware loop.  This patch
prevents this behavior.

This was due to the fact that we always tried to move up the `loopdo`
instruction out of the loop (to keep the previously computed
scheduling within the basic block).  However, doing so would hoist the
second `loopdo` within the first.  Instead of moving up the `loopdo`
instruction we split the head basic block of the loop before its first
instruction, effectively creating an empty basic block before the loop
wherein the `loopdo` instruction will be emitted.  This way, we keep
the scheduling of the loop body, and the `loopdo` instruction ends up
in its own basic block.

gcc/config/kvx:
        * kvx.c (kvx_hwloop_optimize): Fix wrong offset with
          two consecutive hwloops.

Test Plan: CI

Reviewers: bddinechin, O11 gcc

Reviewed By: bddinechin, O11 gcc

Maniphest Tasks: T20580

Differential Revision: https://phab.kalray.eu/D26024