Rewrite the insertion of the hardware loop header
Summary:
gcc/ChangeLog:

2023-01-17  Paul Iannetta  <piannetta@kalray.eu>

	* config/kvx/kvx.c (kvx_hwloop_optimize): rewrite the insertion
	  of the hardware loop header.

Note: The previous change introduced a crash in df-scan.c due to an
improper edition of the CFG.  This time, I generate the simplest
possible CFG such that there is no more error.

```
At first, the loop is a single block with an edge to itself.

      +-------------------+
      |                   |
      |   +-----------+   |
...  -+-->| LOOP BODY |---+-  ...
          +-----------+

This block is then splitted in two.

     +--------------------------------------+
     |                                      |
     |    +-------------+   +-----------+   |
... -+--->| LOOP HEADER |-->| LOOP BODY |---+- ...
          +-------------+   +-----------+

And then, the LOOP HEADER (which contains the loopdo insn), is evicted
from the loop (This is the make_edge and remove_edge).

                          +------------------+
                          |                  |
          +-------------+ v  +-----------+   |
... -+--->| LOOP HEADER |-+->| LOOP BODY |---+- ...
          +-------------+    +-----------+

The idea is to keep the LOOP BODY without modifications so as to not
invalidate any scheduling computation already performed.
```

Test Plan: CI

Reviewers: bddinechin, O11 gcc

Reviewed By: bddinechin, O11 gcc

Differential Revision: https://phab.kalray.eu/D26268