Add a version script to export our vector routines
Summary:
The symbols in libgcc.a are hidden (cf. [1]).  This means that if we
want to dynamically link against a target specific symbol we need to
properly export it through a version script.

This patch collect all our custom symbols and include them in a version
script.  We could have done this earlier and use an earlier version
instead of @GCC_12.0.0.  However, this problem manifested itself
only now (in the dcbu stage when linking), I think we can keep
everything as it is for previous versions and only merges this into
gcc12.

[1]: https://gcc.gnu.org/pipermail/gcc/2012-March/203028.html

libgcc/ChangeLog:

2023-03-17  Paul Iannetta  <piannetta@kalray.eu>

	* libgcc/config/kvx/libgcc-vector.ver: New file.
	* libgcc/config/kvx/t-kvx: Register libgcc-vector.ver into SHLIB_MAPFILES

Test Plan: CI

Reviewers: bddinechin, O11 gcc!

Differential Revision: https://phab.kalray.eu/D27825