[GENERIC] Unconstrain candidates for DO_PREDICATE.
Summary:
In case of SCHED_EBB_PASS with DO_PREDICATE, a REG_DEP_CONTROL dependence
can be paired with a REG_DEP_TRUE dependence by (add_dependence) so that
the EBB scheduler may ignore the REG_DEP_CONTROL dependence in case the
the consumer instruction can be predicated by the test of the jump.
The test performed by (add_dependence) to disable this pairing is
whenever the consumer instruction uses a hard register involved in the
condition for the jump. Replace by a more general condition that disables
pairing only when the register involved in the condition has beed set.

  gcc/gcc/
  * sched-deps.cc: in (add_dependence), replace the test that turns
  `dep_type` to REG_DEP_ANTI from REG_DEP_CONTROL by a call to
  (reg_set_between_p) of the register tested by condition between
  `pro` and `con` instructions; remove unused computation of `uses`;

Test Plan: CI

Reviewers: piannetta, O11 gcc!

Differential Revision: https://phab.kalray.eu/D28877