[GENERIC] Fix problem when scheduling with DO_PREDICATE.
Summary:
Function (schedule_insn) calls (check_clobbered_conditions), which does
nothing unless DO_PREDICATE. If DO_PREDICATE, it checks if the currently
scheduled insn may invalidate the status of the ready and the queued
instructions whose TODO_SPEC is DEP_CONTROL. Invalidation occurs when
the currently scheduled insn sets or clobbers the register used in the
condition of the jump. Indeed the TODO_SPEC DEP_CONTROL instruction
then no longer has the opportunity to be scheduled before the jump by
using predication with the jump condition. So (check_clobbered_conditions)
calls (cond_clobbered_p) on all the ready and the queued instructions
whose TODO_SPEC is DEP_CONTROL, and remove them from the ready or queued
structure if (cond_clobbered_p) returns true.
The problem arises when the jump that produces the control dependency is
already scheduled. In such case the TODO_SPEC DEP_CONTROL instruction
will not have the opportunity to be become ready or queued again.
The fix is to make (cond_clobbered_p) return false in case the
SD_LIST_BACK of the TODO_SPEC DEP_CONTROL instruction is empty.

  gcc/gcc/
  * haifa-sched.cc: in case `insn` passed to (cond_clobbered_p) has
  empty SD_LIST_BACK, restore its INSN_TICK that has been reset by
  (haifa_change_pattern) and return false.

Test Plan: CI

Reviewers: piannetta, O11 gcc!

Differential Revision: https://phab.kalray.eu/D29029