Fix omp_is_initial_device
Summary:
omp_is_initial_device should be 1 on the host and 0 on the offloaded
device.  However, the MPPA can be both used as an accelerating device
and as an host.  In order to not deduplicate the libgomp we introduce a
omp_offline_init function which marks the whole MPPA has an offloading
device.  We may need specialize this function in the future so that each
cluster can be either an host or a device.

libgomp/ChangeLog:

2023-07-24  Paul Iannetta  <piannetta@kalrayinc.com>

	* config/cos/icv-device.c (omp_offload_init): New function.
	(omp_is_initial_device): Check the flag set by
	(omp_offload_init).
	* plugin/plugin-kvx.c (GOMP_OFFLOAD_load_image): Call
	  omp_offload_init at startup.

gcc/ChangeLog:

2023-07-24  Paul Iannetta  <piannetta@kalrayinc.com>

	* config/kvx/mkoffload.cc (process_asm): Create a hidden
	  function to call omp_offload_init from the plugin during the
	  initialization.

Test Plan: CI

Reviewers: tfayard, bddinechin, O11 gcc!

Differential Revision: https://phab.kalray.eu/D30750