Allow non-trivial MMU mappings.
Summary:
The backend needs both virtual and physical addresses in order to talk
to the device.  However, at the boundary between the device and the host
the pointers are opaque (but still have alignment constraints).

At the boundary we use, the device's virtual addresses and we use them
to find back physical addresses.

This commit only works under the hypothesis that virtual addresses are
unique, this is true within a cluster but not true for the whole device.
A future patch will take into account the cluster id when translating
virtual addresses.

libgomp/ChangeLog:

2023-06-20  Paul Iannetta  <piannetta@kalray.eu>

        * plugin/plugin-kvx.c (struct mppa_gomp_buffer): Add a size
          field.
        (struct block_node): New.  Linear datastructure which allows to
        fetch the (vaddr, paddr) pair from a virtual address.
        (find_block): New function. Find the (vaddr, paddr) pair
        associated to paddr.
        (GOMP_OFFLOAD_load_image): Use find_block.
        (GOMP_OFFLOAD_fini_device): Likewise.
        (GOMP_OFFLOAD_alloc): Likewise.
        (GOMP_OFFLOAD_free): Likewise.
        (GOMP_OFFLOAD_dev2host): Likewise.
        (GOMP_OFFLOAD_host2dev): Likewise.
        (GOMP_OFFLOAD_dev2dev): Likewise.
        (GOMP_OFFLOAD_run): Likewise.

Test Plan: CI

Reviewers: tfayard, bddinechin, O11 gcc!

Differential Revision: https://phab.kalray.eu/D30205