ompoffload.kvx: Refactor cluster reservation & async queue access
Summary:
Kernel scheduling now exists as a separate function.
Comments were reworked due to clarity & redundancy issues.
Default kernel queue type is back to RPMSG due to the MMIO mode
having no thread safety.

The async queue mutex was used for two different purposes, so it was
split in two both to fix deadlocks and clarify its usage. There is now a
mutex that controls the access to the queue itself, and a mutex
associated with the condition that wakes up workers whenever a kernel is
available.

Async queue workers need to know which cluster they're supposed to run
their kernel on; so it was added as a field to the kernel. If it isn't
filled upon reaching kvx_kernel_exec, use kvx_schedule_cluster instead.

libgomp/ChangeLog:

2023-07-24  tfayard  <tfayard@kalray.eu>

	* plugin/plugin-kvx.c (struct kernel_info): Change comments &
	variable names.
	(struct async_ctx): Split queue mutex in two, expanded comments.
	(struct agent_info): Remove sysqueue mutexes, add scheduling fields
	(struct kvx_async_worker_params): Added.
	(kvx_schedule_cluster): Add.
	(kvx_async_queue_pop): Refactor.
	(kvx_find_block): Format.
	(kvx_kernel_exec): Move queue type specifics lower, use
	scheduling, explicitly handle all queue types.
	(kvx_enqueue_async_kernel): Refactor, make thread safe.
	(kvx_async_queue_worker): Refactor, use async_worker_params.
	(kvx_init_async_context): Refactor, use pthread_init instead of
	manual memcpy.
	(get_agent_info): Format.
	(kvx_init_agent): Refactor.
	(kvx_init_kernel): Refactor, init new fields.
	(parse_target_attributes): Comment out num_teams & thread_limit.
	(GOMP_OFFLOAD_load_image): Refactor.
	(GOMP_OFFLOAD_fini_device): Same.
	(GOMP_OFFLOAD_can_run): Same.
	(GOMP_OFFLOAD_alloc): Same.
	(GOMP_OFFLOAD_free): Same.

Test Plan: dejagnu integration, CI, manual testing

Reviewers: piannetta, O11 gcc!

Subscribers: tfayard

Differential Revision: https://phab.kalray.eu/D30742