ompoffload.kvx: Keep cluster ID consistent in kernel pipeline
Summary:
Previously, allocations, executions and frees for the same kernel could
be done on different clusters. This doesn't make sense. Kernels and
their command pipeline now run on the same cluster.

A memory space is chosen at the start of kvx_kernel_exec based on
configuration. A cluster ID is derived from the chosen memory space.

If a cluster's SMEM was chosen, we use that same cluster for everything
else. If memory was allocated on global memory, it doesn't matter and
any cluster may be scheduled.

libgomp/ChangeLog:

2023-09-07  tfayard  <tfayard@kalray.eu>

	* plugin/plugin-kvx.c (kvx_kernel_exec): Derive cluster_id from
  alloc_mode.
	(GOMP_OFFLOAD_free): Use alloc_mode.

Test Plan: CI

Reviewers: piannetta, O11 gcc!

Subscribers: tfayard

Differential Revision: https://phab.kalray.eu/D31785