Add XORSIGN* builtin functions.
Summary:
The xorsign builtin is useful to manipulate the sign of a floating-point
number.  GCC has as SPN for this operation, even though it is not exposed
at the front-end level.  Rather, the optimizer matches `r = a *
__builtin_copysign(1.0, b)` when xorsign is in the optab. Add scalar and
vector xorsign<mode> patterns and expose them through kvx builtins.

  gcc/config/kvx/
  * iterators.md: add SCALAR mode iterator.
  * kvx-builtins.cc: implement the XORSIGN* builtins.
  * scalar.md: add expanders for "xorsignhf3", "xorsignsf3", "xorsigndf3".
  * vector.md: refactor the UNSPEC_ANDD and UNSPEC_XORD patterns for use
  in the "xorsign*" patterns; add the "xorsign<mode>" expanders.

Test Plan: CI

Reviewers: piannetta, O11 gcc!

Differential Revision: https://phab.kalray.eu/D32337