kvx: rename XOR* mnemonics to EOR*
Summary:
TCA mnemonics all start with an X, XOR was the only exception.  This
renaming makes it so that only TCA mnemonics start by an X.

On KV3-1 and KV3-2, we install synthetics so that XOR still works.
However, the KV4-1 will not expose a XOR mnemonic.

gcc/ChangeLog:

2023-10-31  Paul Iannetta  <piannetta@kalrayinc.com>

	* config/kvx/extension.md: Rename XOR* mnemonics to EOR*.
	* config/kvx/kvx-builtins.cc (enum kvx_builtin): Likewise.
	(kvx_init_builtins): Likewise.
	(kvx_expand_builtin): Likewise.
	* config/kvx/scalar.md: Likewise.
	* config/kvx/vector.md: Likewise.

Test Plan: CI

Reviewers: bddinechin, O11 gcc!

Differential Revision: https://phab.kalray.eu/D33497