kvx: rename OR* mnemonics to IOR* (cont'd)
TCA instructions start with an X, this introduces ambiguities when it
comes to XOR (Is it the OR on the TCA or the XOR of the core?).  For this
reason, we rename OR to IOR.

The old instructions using OR instead IOR are still supported but are
considered deprecated, and will not be present on Dolomites.

gcc/ChangeLog:

2023-11-07  Paul Iannetta  <piannetta@kalrayinc.com>

	* config/kvx/builtin.md: Rename OR* to IOR*.
	* config/kvx/extension.md: Likewise.
	* config/kvx/scalar.md: Likewise.
	* config/kvx/vector.md: Likewise.