libgomp/posix: remove reclaration of abort
Summary:
This redefinition of `abort' happens within markers which change the
visibility of symbols to `hidden'.  This prevents shared libraries to
link against abort, since the linker first see abort (as an undefined
hidden symbol) in object files which requires it.  When the linker find
the symbol in the libc: if the libc is static (which was the case until
now) everything works fine, however, if the libc is a shared library,
the abort symbol found in the shared library is silently discarded
(because it was marked hidden by the consumer) and the symbol for abort
ends up missing.

Preventing the redefinition avoids abort to be marked as hidden, and
everything works fine again.

libgomp/ChangeLog:

	* config/posix/lock.c: remove abort redefinition.
	* config/posix/mutex.h: Likewise.
	* config/posix/sem.h: Likewise.

Test Plan: Manual

Reviewers: bddinechin, tfayard, O11 gcc

Reviewed By: bddinechin, O11 gcc

Differential Revision: https://phab.kalray.eu/D34966