kvx: rewrite the code handling __builtin_kvx_*
Summary:
The core idea has been taken from the riscv backend.  The idea is to
store the data corresponding to each builtin in a structure.  In
particular, each builtin has an activation predicate and can have a
custom expander if directly calling the pattern from the machine
description is not enough.

	* gcc/config/kvx/kvx-builtins.cc: Refactor.
	* gcc/config/kvx/kvx-builtins.def: New file.
	* gcc/config/kvx/kvx-ftypes.def: New file.
	* gcc/config/kvx/t-kvx: Disable -pedantic.

Test Plan: CI

Reviewers: bddinechin, O11 gcc!

Differential Revision: https://phab.kalray.eu/D35428